Updated UI for list view for Resources and Shifting
