+++ src/Routers/routes/ResourceRoutes.tsx
@@ -1,9 +1,9 @@
 import ResourceDetails from "@/components/Resource/ResourceDetails";
 import { ResourceDetailsUpdate } from "@/components/Resource/ResourceDetailsUpdate";
-import ListView from "@/components/Resource/ListView";
-import BoardView from "@/components/Resource/ResourceBoardView";
 import { Redirect } from "raviger";
 import { AppRoutes } from "../AppRouter";
+import BoardView from "@/components/Resource/ResourceBoard";
+import ListView from "@/components/Resource/ResourceList";
 
 const getDefaultView = () =>
   localStorage.getItem("defaultResourceView") === "list" ? "list" : "board";
