+++ src/Routers/routes/ShiftingRoutes.tsx
@@ -1,8 +1,8 @@
 import { ShiftCreate } from "@/components/Patient/ShiftCreate";
 import ShiftDetails from "@/components/Shifting/ShiftDetails";
 import { ShiftDetailsUpdate } from "@/components/Shifting/ShiftDetailsUpdate";
-import ListView from "@/components/Shifting/ListView";
-import BoardView from "@/components/Shifting/BoardView";
+import ListView from "@/components/Shifting/ShiftingList";
+import BoardView from "@/components/Shifting/ShiftingBoard";
 import { Redirect } from "raviger";
 import { AppRoutes } from "../AppRouter";
