+++ src/components/Resource/ResourceBadges.tsx
