+++ src/components/Resource/ResourceCommons.tsx
