+++ src/components/Resource/ResourceList.tsx
@@ -1,0 +1,137 @@
+import { navigate } from "raviger";
+import ListFilter from "./ResourceFilter";
+import { formatFilter } from "./ResourceCommons";
+import BadgesList from "./ResourceBadges";
+import useFilters from "@/common/hooks/useFilters";
+import { ExportButton } from "@/components/Common/Export";
+import ButtonV2 from "@/components/Common/components/ButtonV2";
+import { useTranslation } from "react-i18next";
+import { AdvancedFilterButton } from "../../CAREUI/interactive/FiltersSlideover";
+import CareIcon from "../../CAREUI/icons/CareIcon";
+import useQuery from "../../Utils/request/useQuery";
+import routes from "../../Redux/api";
+import Page from "@/components/Common/components/Page";
+import SearchInput from "../Form/SearchInput";
+import request from "../../Utils/request/request";
+import Loading from "@/components/Common/Loading";
+import { ResourceModel } from "../Facility/models";
+import ResourceBlock from "./ResourceBlock";
+export default function ListView() {
+  const {
+    qParams,
+    Pagination,
+    FilterBadges,
+    advancedFilter,
+    resultsPerPage,
+    updateQuery,
+  } = useFilters({ cacheBlacklist: ["title"] });
+
+  const { t } = useTranslation();
+
+  const onBoardViewBtnClick = () =>
+    navigate("/resource/board", { query: qParams });
+  const appliedFilters = formatFilter(qParams);
+
+  const { loading, data, refetch } = useQuery(routes.listResourceRequests, {
+    query: formatFilter({
+      ...qParams,
+      offset: (qParams.page ? qParams.page - 1 : 0) * resultsPerPage,
+    }),
+  });
+
+  const showResourceCardList = (data: ResourceModel[]) => {
+    if (data && !data.length) {
+      return (
+        <div className="mt-64 flex flex-1 justify-center text-secondary-600">
+          {t("no_results_found")}
+        </div>
+      );
+    }
+
+    return data.map((resource, i) => (
+      <div key={i} className="rounded-lg border border-secondary-300 bg-white">
+        <ResourceBlock resource={resource} />
+      </div>
+    ));
+  };
+
+  return (
+    <Page
+      title="Resource"
+      hideBack
+      componentRight={
+        <ExportButton
+          action={async () => {
+            const { data } = await request(routes.downloadResourceRequests, {
+              query: { ...appliedFilters, csv: true },
+            });
+            return data ?? null;
+          }}
+          filenamePrefix="resource_requests"
+        />
+      }
+      breadcrumbs={false}
+      options={
+        <>
+          <div className="md:px-4">
+            <SearchInput
+              name="title"
+              value={qParams.title}
+              onChange={(e) => updateQuery({ [e.name]: e.value })}
+              placeholder={t("search_resource")}
+            />
+          </div>
+          <div className="w-32">
+            {/* dummy div to align space as per board view */}
+          </div>
+          <div className="flex w-full flex-col gap-2 lg:w-fit lg:flex-row lg:gap-4">
+            <ButtonV2 className="py-[11px]" onClick={onBoardViewBtnClick}>
+              <CareIcon icon="l-list-ul" className="rotate-90" />
+              {t("board_view")}
+            </ButtonV2>
+
+            <AdvancedFilterButton
+              onClick={() => advancedFilter.setShow(true)}
+            />
+          </div>
+        </>
+      }
+    >
+      <BadgesList {...{ appliedFilters, FilterBadges }} />
+
+      <div className="px-1">
+        {loading ? (
+          <Loading />
+        ) : (
+          <div>
+            <div className="-mb-4 mr-2 mt-4 flex justify-end">
+              <button
+                className="text-xs hover:text-blue-800"
+                onClick={() => refetch()}
+              >
+                <CareIcon
+                  icon="l-refresh"
+                  className="mr-1"
+                  aria-hidden="true"
+                />
+                {t("refresh_list")}
+              </button>
+            </div>
+
+            <div className="mb-5 grid grid-cols-1 gap-4 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 2xl:grid-cols-5">
+              {data?.results && showResourceCardList(data?.results)}
+            </div>
+            <div>
+              <Pagination totalCount={data?.count || 0} />
+            </div>
+          </div>
+        )}
+      </div>
+      <ListFilter
+        {...advancedFilter}
+        showResourceStatus={true}
+        key={window.location.search}
+      />
+    </Page>
+  );
+}
