+++ src/components/Shifting/ShiftDetailsUpdate.tsx
@@ -13,7 +13,7 @@ import {
 import { Cancel, Submit } from "@/components/Common/components/ButtonV2";
 import { navigate, useQueryParams } from "raviger";
 import { useReducer, useState } from "react";
-import { ConsultationModel } from "../Facility/models";
+import { ConsultationModel, ShiftingModel } from "../Facility/models";
 import DischargeModal from "../Facility/DischargeModal";
 import { FacilitySelect } from "@/components/Common/FacilitySelect";
 import { FieldChangeEvent } from "../Form/FormFields/Utils";
@@ -34,7 +34,6 @@ import { LinkedFacilityUsers } from "@/components/Common/UserAutocompleteFormFie
 import { UserBareMinimum } from "../Users/models";
 import useQuery from "../../Utils/request/useQuery";
 import routes from "../../Redux/api";
-import { IShift } from "./models";
 import request from "../../Utils/request/request";
 import { PatientModel } from "../Patient/models";
 import useAuthUser from "@/common/hooks/useAuthUser";
@@ -228,7 +227,7 @@ export const ShiftDetailsUpdate = (props: patientShiftProps) => {
       }
 
       setIsLoading(true);
-      const data: Partial<IShift> = {
+      const data: Partial<ShiftingModel> = {
         origin_facility: state.form.origin_facility_object?.id,
         shifting_approving_facility:
           state.form?.shifting_approving_facility_object?.id,
